+++ addons/custom/l10n_cz_vat_filing/tests/__init__.py
@@ -1,3 +1,4 @@
 from . import test_vat_filing_export
 from . import test_vat_filing_export_wizard
+from . import test_vat_fx_cnb
 from . import test_vat_registry_shield